+++ src/inspect_ai/_eval/evalset.py
@@ -135,6 +135,7 @@ def eval_set(
     log_realtime: bool | None = None,
     log_images: bool | None = None,
     log_model_api: bool | None = None,
+    log_refusals: bool | None = None,
     log_buffer: int | None = None,
     log_shared: bool | int | None = None,
     bundle_dir: str | None = None,
@@ -226,6 +227,7 @@ def eval_set(
         log_images: Log base64 encoded version of images,
             even if specified as a filename or URL (defaults to False)
         log_model_api: Log raw model api requests and responses. Note that error requests/responses are always logged.
+        log_refusals: Log warnings for model refusals.
         log_buffer: Number of samples to buffer before writing log file.
             If not specified, an appropriate default for the format and filesystem is
             chosen (10 for most all cases, 100 for JSON logs on remote filesystems).
@@ -296,6 +298,7 @@ def eval_set(
             log_realtime=log_realtime,
             log_images=log_images,
             log_model_api=log_model_api,
+            log_refusals=log_refusals,
             log_buffer=log_buffer,
             log_shared=log_shared,
             log_header_only=True,
@@ -325,6 +328,7 @@ def eval_set(
         max_subprocesses=max_subprocesses,
         log_level=log_level,
         log_level_transcript=log_level_transcript,
+        log_refusals=log_refusals,
         **kwargs,
     )